style: Api Context, Provider 만들고, interceptor context에서 설정 후 알림 띄우는 코드 이동
