+++ src/api/authApi/_api.ts
@@ -1,11 +1,9 @@
 import axios from 'axios';
 import { API_BASE_URL } from 'api/utils';
 
-const authApi = axios.create({
+export const authApi = axios.create({
   baseURL: API_BASE_URL,
   headers: {
     'Content-Type': 'application/json',
   },
 });
-
-export default authApi;
