+++ src/utils/context/ApiContext.tsx
@@ -1,0 +1,13 @@
+import { createContext } from 'react';
+
+type ApiContextType = {
+  authApi: any;
+  refreshableApi: any;
+};
+
+const ApiContext = createContext<ApiContextType>({
+  authApi: null,
+  refreshableApi: null,
+});
+
+export default ApiContext;
