+++ src/utils/context/index.ts
@@ -1,1 +1,2 @@
 export { default as AlertContext } from './AlertContext';
+export { default as ApiContext } from './ApiContext';
